requirements.txt
appdirs==1.4.3
Django==1.10.6
django-tinymce4-lite==1.3.2
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
six==1.10.0
